Revert back moving of app contents
